Version 0.2.1
Bug fixes:- |<Plug>(ku-choose-action)|: Fix a bug that an error message is displayed unexpectedly, whenever a key which is not associated with any action is pressed.- Fix a bug that the current source name in |the-ku-buffer| is not highlighted correctly in some cases.- Document: Fix a wrong description on the default value of |g:ku_history_reloading_style|.  It is "idle", not "once".- |ku__source|: Fix a bug that it doesn't exist for items which are not completed by ku.- |ku-action-left| and other actions: Fix a bug that they don't close a window or a tab page created by them even if |ku-action-default| is failed.Incompatible changes:- Change the behavior not to take any action if an input pattern is empty and there is no item, because there is nothing to do in such case.- Change the specification on |ku-action|.  Now the function corresponding to an action must return 0 if it succeeds, or a string which represents an error message if it fails.  Old versions don't touch on return values.Misc.:- Refactor the core.