Version 0.2.3
Feature enhancements:- Add |g:ku_personal_runtime| to customize where information such as |ku-input-history| to be saved.- Add |:KuReload| and |ku#reload()| to reload the whole system of ku.Misc. improvements:- Improve an error message.- Add tests, not perfect yet.- |ku-input-history|: Change the directory to save input history.  Use the first directory in the default value of 'runtimepath', instead of the current value of 'runtimepath'.  Because some users customize 'runtimepath', and the first directory in such values may not refer a directory for personal settings.Bug fixes:- Fix a bug that ku list the same sources many times if there are 2 or more source scripts which filenames are the same in 'runtimepath'.- Fix bugs that ku may be confuse if typeahead buffer is not empty at the beginning of a |ku-session|.- |ku#set_the_current_input_pattern()|:  - Fix a wrong description in the document.  - Fix a bug on the treatment of a given string.  If a given string starts with the same character as the prompt, that character is treated as the prompt and not a part of the given string, so it's not easy to set such strings.  - Fix a bug on the automatic completion.  The cursor position may be altered in several cases.See also:- |ku-source-changelog-0.1.2|