Version 0.2.4
Feature enhancements:- |ku-file|: Support to deal with files in archives.  See also |ku-file-changelog-0.1.3| for the details.Misc. improvements:- |filetype-ku|: Modify the default initialization on |the-ku-buffer|.  Old versions don't override existing key mappings.  New versions now override existing key mappings.  It is convenient for starters to use ku without any configuration.  For example, many people customize <Tab> to complete something in  Insert mode.  In such cases, old versions don't override <Tab> to  |<Plug>(ku-choose-an-action)|, but it's very inconvenient to use ku.  All key mappings defined by |ku#default_key_mappings()| are local to  a buffer, so that it's not necessary to avoid overriding.  Thanks to rphillips and other persons for reporting this problem.Bug fixes:- |ku#restart()|: Fix a bug that is caused by a change in the previous version.  As ku#restart() is called, the prompt character will be inserted many times.